+++ tests/testthat/test-tree_generation.R
@@ -34,7 +34,7 @@ test_that("StarTree() works", {
 })
 
 test_that("Random trees are generated correctly", {
-  expect_equal(c(4, 5, 5, 4, 5), RandomTree(3, root = TRUE)$edge[1:5])
+  expect_equal(c(4, 4, 5, 5, 1, 5, 2, 3), RandomTree(3, root = TRUE)$edge[1:8])
   expect_equal(PectinateTree(c('t2', 't3', 't1')), RandomTree(3, root = 't2'))
   expect_equal(c(4, 4, 4), RandomTree(3, root = FALSE)$edge[1:3])
 })
@@ -55,4 +55,20 @@ test_that("EnforceOutgroup() fails nicely", {
                Subtree(Preorder(EnforceOutgroup(letters[1:8], letters[5:6])), 15))
   expect_equal(ape::root(BalancedTree(8), 't1', resolve.root = TRUE),
                EnforceOutgroup(BalancedTree(8), 't1'))
+})
+
+test_that("Random trees drawn from uniform distribution", {
+  nSamples <- 100
+  # Ape's trees are not uniformly distributed:
+  expect_lt(chisq.test(table(vapply(lapply(rep(5, nSamples), ape::rtree),
+                    as.TreeNumber, 1)))$p.value,
+            0.001)
+
+  # Our trees are:
+  expect_gt(chisq.test(table(vapply(lapply(rep(5, nSamples), RandomTree),
+                    as.TreeNumber, 1)))$p.value,
+            0.001)
+
+  expect_false(is.rooted(RandomTree(6, root = FALSE)))
+  expect_true(is.rooted(RandomTree(6, root = TRUE)))
 })